- Implementado Corta & Pega. - Buzones. - Añadidos iconos y otros cambios visuales.

+++ resources/assets/js/components/ExplorerMenuComponent.vue
@@ -2,27 +2,26 @@
 <div>
   <div class="section no-sep">Directorio</div>
   <ul>
-    <li><a class="clickable" v-on:click="newFolder()">Nueva carpeta</a></li>
-    <li><a class="clickable" v-on:click="newProgram()">Nuevo programa</a></li>
+    <li><a class="clickable" v-on:click="newFolder()"><i class="fas fa-folder"></i>Nueva carpeta</a></li>
+    <li><a class="clickable" v-on:click="newProgram()"><i class="fas fa-folder-plus"></i>Nuevo programa</a></li>
   </ul>
   <template v-if="$store.state.explorer.selectionCount > 0">
     <div class="section">Selección ({{ $store.state.explorer.selectionCount }})</div>
     <ul>
-      <li><a class="clickable" v-on:click="deleteSelectedEntries()">Eliminar</a></li>
+      <li v-if="$store.state.explorer.uniqueSelectedEntry && $store.state.explorer.uniqueSelectedEntry.folder"><a class="clickable" v-on:click="editSelectedEntry()"><i class="fas fa-info-circle"></i>Propiedades</a></li>
+      <li v-if="$store.state.explorer.uniqueSelectedEntry && $store.state.explorer.uniqueSelectedEntry.source"><a class="clickable" v-on:click="sendSelectedEntry()"><i class="fas fa-envelope"></i>Enviar</a></li>
+      <li><a class="clickable" v-on:click="cutSelectedEntries()"><i class="fas fa-cut"></i>Cortar</a></li>
+      <li v-if="$store.getters.cuttedEntries.length > 0"><a class="clickable" v-on:click="pasteEntries()"><i class="fas fa-paste"></i>Pegar</a></li>
+      <li><a class="clickable" v-on:click="deleteSelectedEntries()"><i class="fas fa-trash"></i>Eliminar</a></li>
     </ul>
   </template>
-
-  <template v-if="!$store.getters.inRoot && $store.getters.cwd">
-    <div class="section">Buzón <template v-if="$store.getters.cwd.folder.inbox">/{{ $store.getters.cwd.folder.inbox_name }}</template></div>
-    <ul>
-      <li v-if="$store.getters.cwd.folder.inbox == false"><a class="clickable" v-on:click="enableInbox()">Habilitar</a></li>
-      <li v-if="$store.getters.cwd.folder.inbox == true"><a class="clickable" v-on:click="enableInbox()">Deshabilitar</a></li>
-    </ul>
+  <template v-if="$store.getters.cuttedEntries.length > 0 && $store.state.explorer.selectionCount == 0">
+    <div class="section">Ficheros ({{ $store.getters.cuttedEntries.length }})</div>
+    <ul><li><a class="clickable" v-on:click="pasteEntries()"><i class="fas fa-paste"></i>Pegar</a></li></ul>
   </template>
-
   <div class="section">Cuenta</div>
   <ul>
-    <li><a v-on:click="logout()" class="clickable">Cerrar sesión</a></li>
+    <li><a v-on:click="logout()" class="clickable"><i class="fas fa-sign-out-alt"></i>Cerrar sesión</a></li>
   </ul>
 </div>
 </template>
@@ -49,14 +48,28 @@ export default {
       this.$store.dispatch('createSource', { name: name });
     },
 
+    cutSelectedEntries: function() {
+      this.$store.dispatch('cutSelectedEntries');
+    },
+
+    pasteEntries: function () {
+      this.$store.dispatch('pasteEntries');
+    },
+
     deleteSelectedEntries: function () {
-      this.$store.dispatch('deleteSelectedEntries');
+      if (confirm("¿Seguro que quieres eliminar la selección?")) {
+        this.$store.dispatch('deleteSelectedEntries');
+      }
+    },
+
+    editSelectedEntry: function () {
+      this.$store.dispatch('editSelectedEntry');
     },
 
     logout: function () {
       this.$store.dispatch('logout');
       this.$router.push({ name: 'login' });
-    }
+    },
   }
 };
 </script>
